% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{compute_aucs}
\alias{compute_aucs}
\title{Compute AUCs for one or more predictors}
\usage{
compute_aucs(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a vector of AUC values
}
\description{
Compute AUCs with the first component of \code{x} as response and each of the remainings components
as predictors. The first component must be binary (coded as 0, 1).
}
