% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{forward_selection_preds}
\alias{forward_selection_preds}
\title{Select predictors that contribute to the predictive validity of a sum score of the predictors}
\usage{
forward_selection_preds(
  predictors,
  criterion,
  assoc_measure = c("auc", "cor"),
  only_positive = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{predictors}{data frame of predictor variables}

\item{criterion}{a numeric vector representing the criterion (dependent variable)}

\item{assoc_measure}{type of association measure. May be \code{auc} (area under the ROC) or \code{cor} (correlation). For the former, the criterion should be binary (coded 0 for absence, 1 for presence of a feature).}

\item{only_positive}{consider only predictors with nonnegative correlations with the criterion?}

\item{show_progress}{show progress?}
}
\value{
A list containing the components
\itemize{
\item \code{k} (the number of predictors used for the final sum score),
\item \code{assoc} (the association of the final sum score with the criterion),
\item \code{sel_pred_names} (the names of the selected predictors),
\item \code{final_sum_score} (the final sum score of the predictors)
}
}
\description{
Forward selection of predictors that maximize the association of a sum score of the selected predictors with a criterion. All predictors should be coded in the same direction as the criterion, so that the (true) correlation is positive.
}
