% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{mc_crossvalidation}
\alias{mc_crossvalidation}
\title{Monte Carlo Cross-Validation}
\usage{
mc_crossvalidation(x, criterion, n = 100L, assoc_measure = c("auc", "cor"))
}
\arguments{
\item{x}{data frame with predictors}

\item{criterion}{criterion (to be predicted variable)}

\item{n}{number of Monte Carlo runs (i.e., training/validation samples drawn)}

\item{assoc_measure}{type of association measure. May be \code{auc} (area under the ROC) or \code{cor} (correlation). For the former, the criterion should be binary (coded for 0, 1 for presence of a feature).}
}
\value{
A list containing the components
\itemize{
\item \code{assoc} (associations in the training set),
\item \code{assoc_valid} (associations in the validation set), and
\item \code{k} (number of predictors identified in each run).
}

The components are vectors with each value representing the result of one Monte Carlo run.
}
\description{
Stratified Monte Carlo (repeated random sub-sampling) cross-validation for predictor selections with \code{forward_selection_preds}.
}
