% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{mc_crossvalidation_logist}
\alias{mc_crossvalidation_logist}
\title{Monte Carlo Cross-Validation for Predictor Selection with Logistic Regression}
\usage{
mc_crossvalidation_logist(
  x,
  criterion,
  n = 100L,
  only_positive = TRUE,
  plot_auc = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{x}{data frame with predictors}

\item{criterion}{character string specifying the criterion (must be in \code{x})}

\item{n}{number of Monte Carlo runs (i.e., training/validation samples drawn)}

\item{only_positive}{keep only predictors with positive regression coefficients during the initial training run in the model?}

\item{plot_auc}{plot a histogram of the distribution of the AUC for the validation sample?}

\item{show_progress}{show progress?}
}
\value{
A list containing the components
\itemize{
\item \code{auc_train} (AUCs in the training set),
\item \code{auc_valid} (AUCs in the validation set), and
\item \code{k} (number of predictors identified in each run).
}

The components are vectors with each value representing the result of one Monte Carlo run.
}
\description{
Stratified Monte Carlo (repeated random sub-sampling) cross-validation for predictor selections with \link[MASS:stepAIC]{MASS::stepAIC}.
}
\references{
Xu, Q.-S., & Liang, Y.-Z. (2001). Monte Carlo cross validation. \emph{Chemometrics and Intelligent Laboratory Systems}, \emph{56}(1), 1–11. https://doi.org/10.1016/S0169-7439(00)00122-2
}
