% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{mc_crossvalidation_regression}
\alias{mc_crossvalidation_regression}
\title{Monte Carlo Cross-Validation for Predictor Selection with Linear or Logistic Regression}
\usage{
mc_crossvalidation_regression(
  x,
  criterion,
  n = 100L,
  only_positive = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{x}{data frame containing predictors and criterion}

\item{criterion}{character string specifying the criterion (must be in \code{x})}

\item{n}{number of Monte Carlo runs (i.e., training/validation samples drawn)}

\item{only_positive}{keep only predictors with positive regression coefficients in the training run in the model?}

\item{show_progress}{show progress?}
}
\value{
A list containing the components
\itemize{
\item \code{k} (number of predictors identified in each run),
\item \code{method} (logistic or ordinary regression),
\item \code{assoc_train} (associations in the training set),
\item \code{assoc_valid} (associations in the validation set),
\item \code{nonpos_vars_excluded} (number of variables excluded due to nonpositive regression coefficients).
}

The components are vectors with each value representing the result of one Monte Carlo run.
}
\description{
Stratified Monte Carlo (repeated random sub-sampling) cross-validation for predictor selections with \link[stats:step]{stats::step}. The use of ordinary or logistic regression depends on the criterion values. Logistic regression is used when the criterion values are binary and coded with 0 and 1. Otherwise, ordinary regression is used.
}
\references{
Xu, Q.-S., & Liang, Y.-Z. (2001). Monte Carlo cross validation. \emph{Chemometrics and Intelligent Laboratory Systems}, \emph{56}(1), 1–11. https://doi.org/10.1016/S0169-7439(00)00122-2
}
