% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{mc_crossvalidation_sps}
\alias{mc_crossvalidation_sps}
\title{Monte Carlo Cross-Validation for Simple Predictor Selection}
\usage{
mc_crossvalidation_sps(
  x,
  criterion,
  n = 100L,
  assoc_measure = c("auc", "cor"),
  only_positive = TRUE,
  show_progress = FALSE
)
}
\arguments{
\item{x}{data frame containing predictors and criterion}

\item{criterion}{character string specifying the criterion (must be in \code{x})}

\item{n}{number of Monte Carlo runs (i.e., training/validation samples drawn)}

\item{assoc_measure}{type of association measure. May be \code{auc} (area under the ROC) or \code{cor} (correlation). For the former, the criterion should be binary (coded for 0, 1 for presence of a feature).}

\item{only_positive}{consider only predictors with positive correlations with the criterion?}

\item{show_progress}{show progress?}
}
\value{
A list containing the components
\itemize{
\item \code{assoc} (associations in the training set),
\item \code{assoc_valid} (associations in the validation set), and
\item \code{k} (number of predictors identified in each run).
}

The components are vectors with each value representing the result of one Monte Carlo run.
}
\description{
Stratified Monte Carlo (repeated random sub-sampling) cross-validation for predictor selections with \link{sim_pred_sel}.
}
\references{
Xu, Q.-S., & Liang, Y.-Z. (2001). Monte Carlo cross validation. \emph{Chemometrics and Intelligent Laboratory Systems}, \emph{56}(1), 1–11. https://doi.org/10.1016/S0169-7439(00)00122-2
}
