% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{plot.mc_crossvalidation}
\alias{plot.mc_crossvalidation}
\title{Plot Results From a Cross-Validation}
\usage{
\method{plot}{mc_crossvalidation}(x, ...)
}
\arguments{
\item{x}{results from one of the cross-validation functions}

\item{...}{may include \code{binwidth} (binwidth for plot of histograms of the associations, default: 0.01), \code{fill} (fill color for bars, default: "steelblue"), \code{col_mean} (color of lines indicating means, default: "black")}
}
\description{
Plot results from a cross-validation with \link{mc_crossvalidation_sps} or \link{mc_crossvalidation_regression}: a barplot for the distribution of the number of items in the training sets and two histograms for the distributions of the association measures in the training and the validation set.
}
