% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{sim_pred_sel}
\alias{sim_pred_sel}
\title{Simple Predictor Selection}
\usage{
sim_pred_sel(
  x,
  criterion,
  assoc_measure = c("auc", "cor"),
  only_positive = TRUE,
  delta_val = 0,
  show_progress = TRUE,
  show_messages = TRUE
)
}
\arguments{
\item{x}{data frame containing predictors and criterion}

\item{criterion}{character string specifying the criterion (must be in \code{x})}

\item{assoc_measure}{type of association measure. May be \code{auc} (area under the ROC) or \code{cor} (correlation). For the former, the criterion must be binary (coded 0 for absence, 1 for presence of a feature).}

\item{only_positive}{consider only predictors with positive correlations with the criterion?}

\item{delta_val}{minimum validity increment that a chosen predictor must reach}

\item{show_progress}{show progress?}

\item{show_messages}{show messages?}
}
\value{
A list containing the components
\itemize{
\item \code{k} (the number of predictors used for the final sum score),
\item \code{assoc} (the association of the final sum score with the criterion),
\item \code{sel_pred_names} (the names of the selected predictors),
\item \code{final_sum_score} (the final sum score of the selected predictors)
}

If no valid predictors are found NULL is returned.
}
\description{
Find predictors in a set that contribute to the criterion validity of the sum score of the selected predictors.
}
\details{
The search for predictors is done by forward selection. The procedure starts with selecting the predictor with the highest validity (i.e., highest association with the criterion) and then successively adds predictors to the selected set that maximize the \emph{incremental} validity of the sum score.

Association may be measured by the area under the ROC curve (AUC) or by the correlation coefficient. All predictors should be coded in the same direction as the criterion, so that the correlations between the predictors and the criterion are positive. When the association is measured by the AUC, the criterion should be binary, with its values coded 0 or 1 (indicator coding).

Predictors with zero variance are removed from the predictor set before testing incremental validity. By default, predictors with nonpositive correlation with the criterion are also removed in advance. The latter may be included, however, this does normally not matter because predictors with negative correlations with the criterion to rarely contribute to the validity of a sum score. (This would be a kind of suppressor effect.)
}
