% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel.R
\name{forward_selection_preds}
\alias{forward_selection_preds}
\title{Find predictors in a set that contribute to the predictive validity of the sum score of a subset of the predictors}
\usage{
forward_selection_preds(
  predictors,
  criterion,
  assoc_measure = c("auc", "cor"),
  only_positive = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{predictors}{data frame of predictor variables}

\item{criterion}{a numeric vector representing the criterion (dependent variable)}

\item{assoc_measure}{type of association measure. May be \code{auc} (area under the ROC) or \code{cor} (correlation). For the former, the criterion should be binary (coded 0 for absence, 1 for presence of a feature).}

\item{only_positive}{consider only predictors with positive correlations with the criterion?}

\item{show_progress}{show progress?}
}
\value{
A list containing the components
\itemize{
\item \code{k} (the number of predictors used for the final sum score),
\item \code{assoc} (the association of the final sum score with the criterion),
\item \code{sel_pred_names} (the names of the selected predictors),
\item \code{final_sum_score} (the final sum score of the predictors)
}
}
\description{
The search for the best predictors is done by a simple forward selection. The procedure starts with selecting the predictor with the highest validity. Subsequently, predictors with the highest incremental validity are added. In each step, a sum score of all selected preditors is computed whose association with the criterion is assessed. Association may be measured by the area under the ROC curve or by the correlation. All predictors should be coded in the same direction as the criterion, so that the (true) correlation between the predictor and the criterion is positive. By default, predictors with nonpositive correlations with the criterion are eliminated from the predictor set.
}
