% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpredsel-package.R
\docType{package}
\name{simpredsel-package}
\alias{simpredsel}
\alias{simpredsel-package}
\title{simpredsel: Simple Predictor Selection}
\description{
Functions for simple (naive) predictor selection and cross-validation.
}
\author{
\strong{Maintainer}: Michael Hock \email{michael.hock@uni-bamberg.de} (\href{https://orcid.org/0000-0003-1038-3431}{ORCID})

}
\keyword{internal}
